requirements.txt
a==1.0
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cssselect==1.1.0
cssselect2==0.6.0
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.5
django-cors-headers==3.13.0
django-crispy==1.8.1
django-crispy-forms==1.14.0
django-filter==21.1
django-heroku==0.3.1
django-mathfilters==1.0.0
django-qr-code==3.0.0
django-restframework==0.0.1
django-weasyprint==2.1.0
django-wkhtmltopdf==3.4.0
djangorestframework==3.13.1
fonttools==4.33.3
future==0.18.2
gobject==0.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
lxml==4.9.0
oscrypto==1.3.0
pdfkit==1.0.0
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
pydyf==0.1.2
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
pyphen==0.12.0
python-bidi==0.4.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
segno==1.5.2
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.7
zopfli==0.2.1
